ALL STEPS ADDED

+++ hms/src/main/java/com/saad/hms/billing/dto/InvoiceResponseDTO.java
@@ -4,12 +4,14 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
 public class InvoiceResponseDTO {
     private Long id;
-    private Double totalAmount;
-    private Double paidAmount;
+    private BigDecimal totalAmount;
+    private BigDecimal paidAmount;
     private String status;
 }
